Backend

See merge request s06-final/S06P31C202!12

+++ backend/src/main/java/com/ssafy/api/request/user/UserRegisterPostReq.java
@@ -31,5 +31,5 @@ public class UserRegisterPostReq {
 	String email;
 
 	@ApiModelProperty(name = "유저 admin", example = "2")
-	int admin;
+	Long admin;
 }
